web/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
boto3==1.34.88
botocore==1.34.88
certifi==2024.2.2
colorama==0.4.6
crispy-bootstrap5==2024.2
distro==1.9.0
Django==5.0.4
django-crispy-forms==2.1
django-mysql==4.12.0
django-picklefield==3.2
django-q2==1.6.2
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
mysql-connector-python==8.3.0
mysqlclient==2.2.4
numpy==1.26.4
openai==1.23.2
pydantic==2.7.0
pydantic_core==2.18.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
